added average episode duration to UI, added donut chart for watched / all seasons

+++ src/databasemanager.h
@@ -71,7 +71,7 @@ public:
     int seasonCount(int seriesID);
 
     int getWatchedEpisodesDuration();
-    int getAverageWatchedEpisodesDuration();
+    double getAverageWatchedEpisodesDuration();
     int getWatchedEpisodesCount();
     int getAllEpisodesCount();
     int getAllSeriesCount();
